chore: extract error fallback to its own file
